+++ src/src/main/java/com/pkrete/jsip2/messages/requests/SIP2CheckoutRequest.java
@@ -171,7 +171,7 @@ public class SIP2CheckoutRequest extends SIP2CirculationTransactionRequest {
         }
         if (this.useCancel) {
             builder.append("|BI");
-            builder.append(StringUtil.bool2Char(this.useCancel));
+            builder.append(StringUtil.bool2Char(this.cancel));
         }
         builder.append("|");