+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add basic debug and error logging.
 - Make `SIP2MessageRequest.errorDetectionEnabled` an instance variable instead of a class variable.
 - Remove auto increase from `SIP2MessageRequest.sequence` variable. From now on, sequence has to be manually maintained by the client.
+- Fix a bug that caused incorrect `cancel` property value in `SIP2CheckoutRequest` and `SIP2CheckinRequest` messages.
 
 ## 1.1.0 - 2014-06-30
 - Response message fieds with certain values were parsed erroneously and caused exceptions.